Tweeks of the heart rate algorithm to accomidate both iPhone Plus and X
The iPhone X has a yellow tone of color whereas the iPhone Plus is red. The result is that depending upon the device, the calculated heart rate needs to look at either the red or green channel. Additionally, moved to checking the standard deviation on the red channel instead of just a count of red dominant b/c the yellow isn’t recognized as covered. This means that the lens is considered covered by anything that is covering it, but that’s okay for purposes of triggering start.